pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sphinxcontrib-screenshot"
version = "1.1.0"
description = "A Sphinx extension to embed webpage screenshots."
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [
    { name = "Shuhei Iitsuka", email = "tushuhei@gmail.com" }
]
urls = { repository = "https://github.com/tushuhei/sphinxcontrib-screenshot/" }
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
requires-python = ">=3.10"
dependencies = [
    "playwright",
    "sphinx",
    "portpicker",
]

[project.optional-dependencies]
dev = [
    "beautifulsoup4",
    "build",
    "flake8",
    "flake8-pyproject",
    "isort",
    "mypy",
    "Pillow",
    "pre-commit",
    "pytest",
    "pytest-regressions[image]",
    "sphinx[test]",
    "toml",
    "tox",
    "twine",
    "types-beautifulsoup4",
    "types-docutils",
    "types-portpicker",
    "types-Pillow",
    "types-setuptools",
    "user-agents",
    "yapf",
]

doc = [
    "myst-parser",
    "shibuya",
    "sphinx",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["sphinxcontrib*"]

[tool.setuptools.package-data]
"sphinxcontrib.screenshot" = ["static/*"]

[tool.yapf]
based_on_style = "yapf"
column_limit = 79

[tool.flake8]
indent-size = 2

[tool.mypy]
python_version = "3.10"
pretty = true
packages = ["sphinxcontrib"]
explicit_package_bases = true

[tool.tox]
requires = ["tox>=4.19"]
env_list = [
    "py310",
    "py311",
    "py312",
    "py313",
    "py314",
    "style",
]

[tool.tox.env_run_base]
deps = ["-e .[dev]"]
commands = [
    ["pytest", "--showlocals", "--full-trace", "{posargs}"],
]

[tool.tox.env.style]
skip_install = true
commands = [
    ["pre-commit", "run", "--all-files", "--show-diff-on-failure"],
]

[tool.tox.env.doc]
skip_install = true
deps = ["-e .[doc]"]
commands = [
    ["sphinx-build", "--builder", "html", "doc", "build/sphinx/html"],
]
